requirements.txt
spotipy==2.23.0
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
flask==3.0.0
python-dotenv==1.0.0
requests==2.31.0
plotly==5.17.0
dash==2.14.2
dash-bootstrap-components==1.5.0
gunicorn==21.2.0 